requirements.txt
fastapi 
uvicorn 
python-multipart 
aiofiles 
openai
azure-identity